Enhanced create RpcClient and FlumeSink
