+++ flink-connector-flume/src/main/java/org/apache/flink/streaming/connectors/flume/FlumeUtils.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.connectors.flume.v2;
+package org.apache.flink.streaming.connectors.flume;
 
 import org.apache.flume.api.RpcClient;
 import org.apache.flume.api.RpcClientConfigurationConstants;
