--- flink-connector-flume/src/test/java/org/apache/flink/streaming/connectors/flume/v1/FlumeRpcClientTest.java
@@ -1,68 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.flink.streaming.connectors.flume.v1;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-@DockerTest
-public class FlumeRpcClientTest {
-
-    public FlumeRpcClient createGoodClient() {
-        return new FlumeRpcClient("172.25.0.3", 44444);
-    }
-
-    @Test
-    public void testInitClientMustFail() {
-        FlumeRpcClient client = new FlumeRpcClient("172.25.0.3", 44445);
-        Assertions.assertThrows(RuntimeException.class, () -> client.init(), "client start");
-    }
-
-    @Test
-    public void testSendStringData() {
-        FlumeRpcClient client = createGoodClient();
-        boolean init = client.init();
-        Assertions.assertTrue(init, "client not start");
-
-        boolean send = client.sendData("xpto");
-        Assertions.assertTrue(send, "data not send");
-
-    }
-
-    @Test
-    public void testSendBytesData() {
-        FlumeRpcClient client = createGoodClient();
-        boolean init = client.init();
-        Assertions.assertTrue(init, "client not start");
-
-        boolean send = client.sendData("xpto".getBytes());
-        Assertions.assertTrue(send, "data not send");
-
-    }
-
-    @Test
-    public void testSendDataWhenConnectionClosed() {
-        FlumeRpcClient client = createGoodClient();
-        boolean init = client.init();
-        Assertions.assertTrue(init, "client not start");
-        client.close();
-
-        boolean send = client.sendData("xpto");
-        Assertions.assertTrue(send, "data not send");
-
-    }
-}
